Added the rest of the timing patch (#2605900) -- _this_ time it works (: Tested with DoA, no problems.  The last version of the patch reset [MS]SH2->cycles before calling FRTExec(), so the FRC almost never got incremented.
Question: Wouldn't it be more correct to add the nominal cycles per call
(the "cycles" parameter) rather than the value of CurrentSH2->cycles (which
may be 1 or 2 cycles over depending on the previous instruction?)  E.g. at
170 cycles per call:
    (Call 1) 171 cycles executed, MSH2->cycles = 171
             FRC takes 171 cycles
             MSH2->cycles -= 170 ==> value is now 1
    (Call 2) 169 cycles executed, MSH2->cycles = 170
             FRC takes 170 cycles
             MSH2->cycles -= 170 ==> value is now 0
So after 2 calls of 170 cycles, the SH2 core has executed 340 cycles
exactly, but the FRC has taken 341 cycles, which seems wrong.


git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1996 d7be59e3-070e-4b9a-85a1-4646befae34d