Update einc immediately when changing registers it depends on in scsp.c
Make einc a pointer to the respective increment it has been set to so that
it updates when these change, and make sure to change the envelope
increments where necessary.

This introduces a slightly incompatibility with save states, and may result
in odd sound when switching formats until things settle.

git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@2831 d7be59e3-070e-4b9a-85a1-4646befae34d