Adding a config.h.in file...
Ok, this change is somehow related to the Android port, but may
be interesting for any other port that want to build the Yabause
library with CMake but use something else to build itself.
This creates a config.h file in the build directory, this file
should be filled with #define matching the -D used when building
the library.
For now, I only moved SH2_DYNAREC to this file as that's what I
need for the Android, but I should move others soon...

git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@2872 d7be59e3-070e-4b9a-85a1-4646befae34d