New vdp2 "debug texture" function.
Some time ago, when I added the titan library, I broke a feature
used by the windows port: debugging the vdp2 textures.
This worked by calling the software renderer's Vdp2DrawScroll
function on a user-defined texture buffer. But when I added the
titan library, I also removed the texture parameter in Vdp2DrawScroll.

The software renderer now provide a function to draw one of the
vdp2 screen and the "debug texture" function use it.

The windows port now should be really broken as I have changed
the function signature... but I'll fix it in the following commits.

git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@2895 d7be59e3-070e-4b9a-85a1-4646befae34d