New OSD system.
This is moving the OSD stuff out of the video cores
into a new "core system". The OSD cores only needs one function.
For now, there's only two OSD cores: dummy and glut.

Instead of breaking about everything, I've tried to make
this change as transparent as possible:
- by default, yabause provides an OSD core list, but you can
  override it at compile time;
- I added a number of backward compatibility functions.

That means that gtk, qt, windows and android ports still compile
without a change. On the other hand I broke the GCD renderer and
thus, compilation of the cocoa port. I'm not sure of the status
of the other ports...

As for *why* I've done that change... I want to make it possible
to choose the OSD core: for instance, glut is not available when
yabause is built in "full software" mode... there's also ports
living outside the yabause repository that hack the yabause
library to include their OSD system... I hope it'll make things
easier for them (yeah, I'm talking about you wii and xbox ports)

git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@2907 d7be59e3-070e-4b9a-85a1-4646befae34d