Improvements to the new OSD system.
There's two things included in this commit:
- a new OSD core;
- an update to the OSD api.

The new OSD core is a full software one, with a bitmap font
included. The font may be a bit small and is not complete yet...
but it's enough to display the FPS.

The new api introduces a new function: OSDUseBuffer that tells
if the OSD core needs a buffer to draw into or if it will
render to OpenGL.
It also add parameters to OSDDisplayMessages to provide a
buffer and its size.

The default OSD core is now Glut if it's found and the
software one otherwise. The OpenGL renderer is not able to use
the software OSD yet.

git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@2957 d7be59e3-070e-4b9a-85a1-4646befae34d