Ports now have control over the OSD cores.
When compiled with the new YAB_PORT_OSD option, the emulation library
doesn't provide a OSDCoreList anymore and let ports choose their osd
core with the osdcoretype configuration in YabauseInit.
Of course, this means that trying to compile ports that don't handle
YAB_PORT_OSD will fail because of missing OSDCoreList symbol.

git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@2963 d7be59e3-070e-4b9a-85a1-4646befae34d