First partial merge of devmiyax work
This adds a new sound core to the Android port using the OpenSL library.
The new core is used as the new default one.

It can be disabled at compilation time by setting YAB_WANT_OPENSL to OFF,
so people who has really old Android device can still use Yabause...

Ok, that's only 0.3% of the Android devices at the time of this commit,
but I'm one of those 0.3% ;)