Support for different Android platforms
This change make it possible to select an Android Platform
in the different android cmake files.

It changes the platform used to compile the Yabause library
and set APP_PLATFORM to the same value for ndk-build.

Platform is set to android-19 for armeabi, making it possible
to run Yabause on really old platforms, like android-8.